classification/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
_tflow_select=2.3.0=eigen
absl-py=1.3.0=py38haa95532_0
aiohttp=3.8.3=py38h2bbff1b_0
aiosignal=1.2.0=pyhd3eb1b0_0
antlr4-python3-runtime=4.9.3=pypi_0
argon2-cffi=21.3.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
astor=0.8.1=py38haa95532_0
asttokens=2.0.7=pypi_0
astunparse=1.6.3=py_0
async-timeout=4.0.2=py38haa95532_0
attrs=22.1.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.11.1=pypi_0
blas=1.0=mkl
bleach=5.0.1=pypi_0
blinker=1.4=py38haa95532_0
bottleneck=1.3.5=py38h080aedc_0
brotli=1.0.9=ha925a31_2
brotlipy=0.7.0=py38h2bbff1b_1003
ca-certificates=2023.01.10=haa95532_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.12.7=py38haa95532_0
cffi=1.15.0=py38h2bbff1b_1
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py38haa95532_0
colorama=0.4.5=py38haa95532_0
configargparse=1.4.1=pypi_0
cryptography=36.0.0=py38h21b164f_0
cudatoolkit=11.3.1=h59b6b97_2
cvxopt=1.3.0=py38he924ae6_1
cycler=0.11.0=pyhd3eb1b0_0
debugpy=1.6.2=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
dsdp=5.8=h6e01ec9_1203
easydict=1.9=py_0
eigen=3.3.7=h59b6b97_1
entrypoints=0.4=pypi_0
exceptiongroup=1.1.0=pyhd8ed1ab_0
executing=0.9.1=pypi_0
fastjsonschema=2.16.1=pypi_0
fftw=3.3.10=nompi_h52fa85e_103
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.10.4=hd328e21_0
frozenlist=1.3.3=py38h2bbff1b_0
gast=0.4.0=pyhd3eb1b0_0
giflib=5.2.1=h8cc25b3_3
glib=2.69.1=h5dc1a3c_2
glpk=4.65=h8ffe710_1004
google-auth=2.6.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.4=pyhd3eb1b0_0
google-pasta=0.2.0=pyhd3eb1b0_0
grpcio=1.42.0=py38hc60d5dd_0
gsl=2.7=hdfb1a43_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
h5py=2.10.0=py38h5e291fa_0
hdf5=1.10.4=h7ebc959_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha925a31_3
idna=3.3=pyhd3eb1b0_0
importlib-metadata=4.11.3=py38haa95532_0
importlib-resources=5.9.0=pypi_0
iniconfig=2.0.0=pyhd8ed1ab_0
intel-openmp=2021.4.0=haa95532_3556
ipykernel=6.15.1=pypi_0
ipython=8.4.0=pypi_0
ipython-genutils=0.2.0=pypi_0
ipywidgets=7.7.1=pypi_0
jedi=0.18.1=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.1.0=pyhd3eb1b0_0
jpeg=9e=h2bbff1b_1
jsonschema=4.9.1=pypi_0
jupyter-client=7.3.4=pypi_0
jupyter-core=4.11.1=pypi_0
jupyter-http-over-ws=0.0.8=pypi_0
jupyterlab-pygments=0.2.2=pypi_0
jupyterlab-widgets=1.1.1=pypi_0
keras-applications=1.0.8=py_1
keras-preprocessing=1.1.2=pyhd3eb1b0_0
kiwisolver=1.3.2=py38hd77b12b_0
lerc=3.0=hd77b12b_0
libblas=3.9.0=1_h8933c1f_netlib
libcblas=3.9.0=5_hd5c7e75_netlib
libclang=12.0.0=default_h627e005_2
libdeflate=1.17=h2bbff1b_0
libffi=3.4.2=hd77b12b_6
libiconv=1.16=h2bbff1b_2
liblapack=3.9.0=5_hd5c7e75_netlib
libogg=1.3.5=h2bbff1b_1
libpng=1.6.37=h2a8f88b_0
libprotobuf=3.20.3=h23ce68f_0
libtiff=4.5.0=h8a3f274_0
libuv=1.40.0=he774522_0
libvorbis=1.3.7=he774522_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
libxml2=2.9.14=h0ad7f3c_0
libxslt=1.1.35=h2bbff1b_0
lz4-c=1.9.3=h2bbff1b_1
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.4.1=py38haa95532_0
markupsafe=2.1.1=py38h2bbff1b_0
matplotlib=3.5.1=py38haa95532_1
matplotlib-base=3.5.1=py38hd77b12b_1
matplotlib-inline=0.1.3=pypi_0
mistune=0.8.4=pypi_0
mkl=2021.4.0=haa95532_640
mkl-service=2.4.0=py38h2bbff1b_0
mkl_fft=1.3.1=py38h277e83a_0
mkl_random=1.2.2=py38hf11a4ad_0
msys2-conda-epoch=20160418=1
multidict=6.0.2=py38h2bbff1b_0
munkres=1.1.4=py_0
nbclient=0.6.6=pypi_0
nbconvert=6.5.0=pypi_0
nbformat=5.4.0=pypi_0
nest-asyncio=1.5.5=pypi_0
ninja=1.10.2=py38h559b2a2_3
notebook=6.4.12=pypi_0
numexpr=2.8.3=py38hb80d3ca_0
numpy=1.21.5=py38ha4e8547_0
numpy-base=1.21.5=py38hc2deb75_0
nvidia-apex=0.1=py38hea2869c_4
oauthlib=3.2.1=py38haa95532_0
omegaconf=2.3.0=pypi_0
opencv=4.6.0=py38h104de81_2
opencv-contrib-python=4.7.0.72=pypi_0
openssl=1.1.1t=h2bbff1b_0
opt_einsum=3.3.0=pyhd3eb1b0_1
packaging=21.3=pyhd3eb1b0_0
pandas=1.5.3=py38hf11a4ad_0
pandocfilters=1.5.0=pypi_0
parso=0.8.3=pypi_0
pcre=8.45=hd77b12b_0
pickleshare=0.7.5=pypi_0
pillow=9.0.0=pypi_0
pip=21.2.2=py38haa95532_0
pkgutil-resolve-name=1.3.10=pypi_0
pluggy=1.0.0=pyhd8ed1ab_5
ply=3.11=py38_0
prometheus-client=0.14.1=pypi_0
prompt-toolkit=3.0.30=pypi_0
protobuf=3.20.3=py38hd77b12b_0
psutil=5.9.1=pypi_0
pure-eval=0.2.2=pypi_0
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.12.0=pypi_0
pyjwt=2.4.0=py38haa95532_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.4=pyhd3eb1b0_0
pyqt=5.15.7=py38hd77b12b_0
pyqt5-sip=12.11.0=py38hd77b12b_0
pyreadline=2.1=py38_1
pyrsistent=0.18.1=pypi_0
pysocks=1.7.1=py38haa95532_0
pytest=7.2.1=pyhd8ed1ab_0
python=3.8.12=h6244533_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytorch=1.10.1=py3.8_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py38haa95532_0
pywin32=304=pypi_0
pywinpty=2.0.7=pypi_0
pyyaml=6.0=py38h294d835_4
pyzmq=23.2.0=pypi_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h3ad3cdb_4
randaugment=1.0.2=pypi_0
requests=2.27.1=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-learn=1.0.2=py38hf11a4ad_1
scipy=1.9.1=pypi_0
seaborn=0.12.2=py38haa95532_0
send2trash=1.8.0=pypi_0
setuptools=58.0.4=py38haa95532_0
sip=6.6.2=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
soupsieve=2.3.2.post1=pypi_0
sqlite=3.41.1=h2bbff1b_0
stack-data=0.3.0=pypi_0
tensorboard=2.10.0=py38haa95532_0
tensorboard-data-server=0.6.1=py38haa95532_0
tensorboard-plugin-wit=1.8.1=py38haa95532_0
tensorflow=2.3.0=mkl_py38h8c0d9a2_0
tensorflow-base=2.3.0=eigen_py38h75a453f_0
tensorflow-estimator=2.6.0=pyh7b7c402_0
termcolor=2.1.0=py38haa95532_0
terminado=0.15.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2023.2.28=pypi_0
tinycss2=1.1.1=pypi_0
tk=8.6.11=h2bbff1b_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=pyhd8ed1ab_0
torchaudio=0.10.1=py38_cu113
torchvision=0.11.2=py38_cu113
tornado=6.1=py38h2bbff1b_0
tqdm=4.64.0=py38haa95532_0
traitlets=5.3.0=pypi_0
typing_extensions=3.10.0.2=pyh06a4308_0
urllib3=1.26.8=pyhd3eb1b0_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
wcwidth=0.2.5=pypi_0
webencodings=0.5.1=pypi_0
werkzeug=2.2.2=py38haa95532_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=3.6.1=pypi_0
win_inet_pton=1.1.0=py38haa95532_0
wincertstore=0.2=py38haa95532_2
wrapt=1.14.1=py38h2bbff1b_0
xz=5.2.5=h62dcd97_0
yaml=0.2.5=h8ffe710_2
yarl=1.8.1=py38h2bbff1b_0
zipp=3.8.1=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
